[metaparser.cpp] Fixed integer overflow for long recordings recordings over 6h37 (2^31 pts) had negative m_length in meta file resulting in wrong progress bar and wrong end time in EMC - thx Fischreiher

+++ lib/gui/epositiongauge.cpp
@@ -203,7 +203,7 @@ int ePositionGauge::scale(const pts_t &val)
 
 	int width = size().width();
 
-	return width * val / m_length;
+	return (int)(width * val / m_length);
 }
 
 void ePositionGauge::setForegroundColor(const gRGB &col)
